requirements.txt
python==3.10
streamlit==1.50.0
pandas==2.2.3
plotly==6.3.1
seaborn==0.13.2
matplotlib==3.9.4
scikit-learn==1.6.1
kaleido==0.2.1
biopython==1.84
pysam==0.22.1